move dyson subcomponents to the main component module
