+++ homeassistant/components/dyson/sensor.py
