+++ tests/components/dyson/test_climate.py
@@ -4,10 +4,11 @@ from unittest import mock
 
 from libpurecool.const import (FocusMode, HeatMode,
                                HeatState, HeatTarget, TiltState)
-from libpurecool.dyson_pure_state import DysonPureHotCoolState
 from libpurecool.dyson_pure_hotcool_link import DysonPureHotCoolLink
-from homeassistant.components.climate import dyson
+from libpurecool.dyson_pure_state import DysonPureHotCoolState
+
 from homeassistant.components import dyson as dyson_parent
+from homeassistant.components.dyson import climate as dyson
 from homeassistant.const import TEMP_CELSIUS, ATTR_TEMPERATURE
 from homeassistant.setup import setup_component
 from tests.common import get_test_home_assistant
