+++ tests/components/dyson/test_sensor.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 from libpurecool.dyson_pure_cool_link import DysonPureCoolLink
 
-from homeassistant.components.sensor import dyson
+from homeassistant.components.dyson import sensor as dyson
 from homeassistant.const import TEMP_CELSIUS, TEMP_FAHRENHEIT, \
     STATE_OFF
 from tests.common import get_test_home_assistant
